Use DependencyExtractor in TupleAnalyzer and RelationPlanner
