+++ presto-main/src/main/java/com/facebook/presto/sql/planner/DependencyExtractor.java
@@ -15,6 +15,7 @@ package com.facebook.presto.sql.planner;
 
 import com.facebook.presto.sql.tree.DefaultExpressionTraversalVisitor;
 import com.facebook.presto.sql.tree.Expression;
+import com.facebook.presto.sql.tree.QualifiedName;
 import com.facebook.presto.sql.tree.QualifiedNameReference;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
@@ -22,7 +23,6 @@ import com.google.common.collect.ImmutableSet;
 import java.util.List;
 import java.util.Set;
 
-// TODO: a similar class exists in (TupleAnalyzer.DependencyExtractor)
 public final class DependencyExtractor
 {
     private DependencyExtractor() {}
@@ -44,17 +44,36 @@ public final class DependencyExtractor
     public static List<Symbol> extractAll(Expression expression)
     {
         ImmutableList.Builder<Symbol> builder = ImmutableList.builder();
-        new Visitor().process(expression, builder);
+        new SymbolBuilderVisitor().process(expression, builder);
         return builder.build();
     }
 
-    private static class Visitor
+    // to extract qualified name with prefix
+    public static Set<QualifiedName> extractNames(Expression expression)
+    {
+        ImmutableSet.Builder<QualifiedName> builder = ImmutableSet.builder();
+        new QualifiedNameBuilderVisitor().process(expression, builder);
+        return builder.build();
+    }
+
+    private static class SymbolBuilderVisitor
             extends DefaultExpressionTraversalVisitor<Void, ImmutableList.Builder<Symbol>>
     {
         @Override
         protected Void visitQualifiedNameReference(QualifiedNameReference node, ImmutableList.Builder<Symbol> builder)
         {
             builder.add(Symbol.fromQualifiedName(node.getName()));
+            return null;
+        }
+    }
+
+    private static class QualifiedNameBuilderVisitor
+            extends DefaultExpressionTraversalVisitor<Void, ImmutableSet.Builder<QualifiedName>>
+    {
+        @Override
+        protected Void visitQualifiedNameReference(QualifiedNameReference node, ImmutableSet.Builder<QualifiedName> builder)
+        {
+            builder.add(node.getName());
             return null;
         }
     }
